fix: export module augumetation
